Fix for #1806
View3d windows have 'unlock' option and 'localview', which was also
supported on render (F12 with mouse in window) to render only what is
visible there.
This didn't work very nice, with even code in render module to read from
interface variable. Removed that from render, and coded the exception
where it belongs, in renderwin.c