- readfile incorrectly increased user count for Lattice & Curve    modifier objects  - make unlink_object clear Lattice & Curve Modifier object references  - add expand_modifiers for liblinking
There really needs to be a more consistent and unified way of dealing
with datablock references between objects. It should be possible to
make this generic so that lib_link, expand_, DEP graph, unlink,
oops and maybe outliner can use a single API. The code to deal with
this is too bulky and error prone at the moment.